# Improve performance when entering large menus