--- a/Celeste.Mod.mm/Patches/TextMenu.cs
+++ b/Celeste.Mod.mm/Patches/TextMenu.cs
@@ -16,12 +16,16 @@ namespace Celeste {
         private List<Item> items;
 
         private bool batchMode;
+        private float width;
+        private float leftColumnWidth;
+        private float rightColumnWidth;
+        private float height;
+        private bool recalculatingSizeInBatchMode;
 
         /// <summary>
         /// The items contained in this menu.
         /// </summary>
         public List<Item> Items => items;
-
 
         /// <summary>
         /// When a menu is in batch mode, adding / removing items will not recalculate its size to improve performance.
@@ -35,6 +39,58 @@ namespace Celeste {
                     RecalculateSize();
                 }
             }
+        }
+
+        /// <inheritdoc cref="TextMenu.Width"/>
+        public new float Width {
+            [MonoModReplace]
+            get {
+                if (batchMode && !recalculatingSizeInBatchMode) {
+                    RecalculateSize();
+                }
+                return width;
+            }
+            [MonoModReplace]
+            private set => width = value;
+        }
+
+        /// <inheritdoc cref="TextMenu.Height"/>
+        public new float Height {
+            [MonoModReplace]
+            get {
+                if (batchMode && !recalculatingSizeInBatchMode) {
+                    RecalculateSize();
+                }
+                return height;
+            }
+            [MonoModReplace]
+            private set => height = value;
+        }
+
+        /// <inheritdoc cref="TextMenu.LeftColumnWidth"/>
+        public new float LeftColumnWidth {
+            [MonoModReplace]
+            get {
+                if (batchMode && !recalculatingSizeInBatchMode) {
+                    RecalculateSize();
+                }
+                return leftColumnWidth;
+            }
+            [MonoModReplace]
+            private set => leftColumnWidth = value;
+        }
+
+        /// <inheritdoc cref="TextMenu.RightColumnWidth"/>
+        public new float RightColumnWidth {
+            [MonoModReplace]
+            get {
+                if (batchMode && !recalculatingSizeInBatchMode) {
+                    RecalculateSize();
+                }
+                return rightColumnWidth;
+            }
+            [MonoModReplace]
+            private set => rightColumnWidth = value;
         }
 
         // Basically the same as Add(), but with an index parameter.
@@ -188,6 +244,20 @@ namespace Celeste {
             }
             item.Added();
             return this;
+        }
+
+#pragma warning disable CS0626 // extern method with no attribute
+        public extern void orig_RecalculateSize();
+#pragma warning restore CS0626 // extern method with no attribute
+
+        public new void RecalculateSize() {
+            if (BatchMode) {
+                recalculatingSizeInBatchMode = true;
+            }
+            orig_RecalculateSize();
+            if (BatchMode) {
+                recalculatingSizeInBatchMode = false;
+            }
         }
 
         public class patch_LanguageButton : LanguageButton {