### Issue

If players install a lot of mods, the game may freeze for a short time when they enter a menu with a lot of items (e.g. mod options, map search, map list).

### Reason

The game calls `TextMenu.RecalculateSize()` and calculates all items' width and height every time an item is added or removed. It is unnecessary because this happened in one frame and the calculation is only needed after all operations are done.

### Solution

This patch adds a batch mode to the text menu. When it is enabled, recalculating menu size is skipped when adding and removing menu items, which can significantly increase performance when players enter a large menu.

If `Width` / `Height` / `LeftColumnWidth` / `RightColumnWidth` is accessed in batch mode, `TextMenu.RecalculateSize()` will be called to make sure the result is correct. If `TextMenu.RecalculateSize()` is called in batch mode, `recalculatingSizeInBatchMode` will be set to `true` to prevent infinite recursive calls.

The patch also optimizes `Emoji.Apply(string)` by using two pointers to iterate the string only once and replace the emojis, to also improve performance in large menus, since `TextMenu.RecalculateSize()` is called every time the menu renders. Although there aren't many mods that register emojis, CelesteNet used emojis to display avatars, so if someone has been online for hours, about 50 avatars can be registered as emojis, causing lags on menus.

### Result

Suppose that we have a text menu with `n` items, `a` is an array of length `n`, containing each item's text length, and there are `m` emojis registered. The time complexity before and after the patch is:

|        | Best                | Average             | Worst                 |
| ------ | ------------------- | ------------------- | --------------------- |
| Before | `O(n * m * sum(a))` | `O(n * m * sum(a))` | `O(n^2 * m * max(a))` |
| After  | `O(sum(a))`         | `O(sum(a))`         | `O(sum(a))`           |

Call count and time comparison when entering map list with 20 level sets and 400 maps installed, all maps' name is 20 characters long without emojis, and there are 30 emojis registered: (Note that call time is much longer than in real game due to the profiler)

Before:

![before](https://user-images.githubusercontent.com/7558201/121803887-1fffd800-cc76-11eb-8327-8260e2838a77.png)

After:

![after](https://user-images.githubusercontent.com/7558201/121803891-23935f00-cc76-11eb-8b0f-b5c7d79b4d2a.png)

`Emoji.Apply()` benchmark result before and after optimizing ([code](https://www.pastery.net/mgajdh/)):

```text
BenchmarkDotNet=v0.10.3.0, OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-7700HQ CPU 2.80GHz, ProcessorCount=8
Frequency=2742190 Hz, Resolution=364.6720 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.8.4010.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.8.4010.0
```

| Method | Emoji Count | Emoji in Text |               Mean |          StdErr |          StdDev | Scaled |
| ------ | ----------: | ------------: | -----------------: | --------------: | --------------: | -----: |
| Before |          10 |            10 |      3,095.7799 ns |      24.8549 ns |      96.2625 ns |   1.00 |
| After  |          10 |            10 |        912.1202 ns |       9.0746 ns |      48.0184 ns |   0.29 |
| Before |          10 |           100 |     27,512.6470 ns |     225.4702 ns |     873.2423 ns |   1.00 |
| After  |          10 |           100 |      9,595.4766 ns |      88.8038 ns |     486.3986 ns |   0.35 |
| Before |          10 |          1000 |    261,457.5667 ns |   2,535.8593 ns |  11,340.7077 ns |   1.00 |
| After  |          10 |          1000 |     95,803.9844 ns |     772.2700 ns |   2,990.9888 ns |   0.37 |
| Before |         100 |            10 |     12,256.1060 ns |      87.8604 ns |     340.2819 ns |   1.00 |
| After  |         100 |            10 |        865.7724 ns |       8.3928 ns |      41.1159 ns |   0.07 |
| Before |         100 |           100 |    190,369.1403 ns |   1,087.8736 ns |   4,213.3164 ns |   1.00 |
| After  |         100 |           100 |      8,615.6526 ns |      82.9249 ns |     388.9523 ns |   0.05 |
| Before |         100 |          1000 |  2,671,179.0389 ns |  31,067.6264 ns | 194,017.2648 ns |   1.00 |
| After  |         100 |          1000 |     97,382.9423 ns |     808.6068 ns |   3,025.5296 ns |   0.04 |
| Before |        1000 |            10 |    105,540.6256 ns |   1,045.1672 ns |   5,628.3978 ns |   1.00 |
| After  |        1000 |            10 |        943.0521 ns |       9.2539 ns |      47.1859 ns |   0.01 |
| Before |        1000 |           100 |    534,928.4297 ns |   4,262.8933 ns |  16,510.1148 ns |   1.00 |
| After  |        1000 |           100 |      9,600.3199 ns |      64.1450 ns |     248.4326 ns |   0.02 |
| Before |        1000 |          1000 | 16,723,940.9119 ns | 161,419.3316 ns | 665,548.9541 ns |   1.00 |
| After  |        1000 |          1000 |     84,667.3785 ns |     827.3982 ns |   3,309.5929 ns |   0.01 |